Serve metrics back in main. Updated toml
