+++ Gopkg.toml
@@ -4,13 +4,16 @@
 
 [[constraint]]
   name = "github.com/stretchr/testify"
-
-[[constraint]]
-  name = "k8s.io/api"
-
-[[constraint]]
-  name = "k8s.io/apimachinery"
+  branch = "master"
 
 [[constraint]]
   name = "k8s.io/client-go"
   version = "5.0.1"
+
+[[constraint]]
+  name = "github.com/prometheus/client_golang/prometheus/promhttp"
+  version = "0.8.0"
+
+[[constraint]]
+  name = "github.com/prometheus/client_golang/prometheus"
+  version = "0.8.0"
