+++ cmd/kapprover/main.go
@@ -2,12 +2,15 @@ package main
 
 import (
 	"flag"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/proofpoint/kapprover/inspectors"
 	"github.com/proofpoint/kapprover/kapprover"
 	log "github.com/sirupsen/logrus"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
+	"net/http"
+	"strconv"
 	"time"
 
 	_ "github.com/proofpoint/kapprover/inspectors/altnamesforpod"
@@ -45,8 +48,19 @@ func main() {
 		return
 	}
 
-	kapprover.ServePrometheusMetrics(metricsPort)
+	go serveMetrics(metricsPort)
 	kapprover.HandleRequests(filters, deniers, warners, *deleteAfter, client)
+}
+
+func serveMetrics(port int) {
+	http.HandleFunc("/healthz", func(w http.ResponseWriter, _ *http.Request) {
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte("OK"))
+	})
+
+	http.Handle("/metrics", promhttp.Handler())
+
+	log.Fatal(http.ListenAndServe(":"+strconv.Itoa(port), nil))
 }
 
 func newClient(kubeconfigPath string) (*kubernetes.Clientset, error) {
