+++ kapprover/kapprover.go
@@ -3,7 +3,6 @@ package kapprover
 import (
 	"fmt"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/proofpoint/kapprover/inspectors"
 	log "github.com/sirupsen/logrus"
 	certificates "k8s.io/api/certificates/v1beta1"
@@ -12,8 +11,6 @@ import (
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/cache"
-	"net/http"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -58,14 +55,9 @@ var (
 			Name: "kapprover_requests_error",
 			Help: "Number of requests encountering an error.",
 		},
-		[]string{"message"},
+		[]string{"reason"},
 	)
 )
-
-func ServePrometheusMetrics(port int) {
-	registerPrometheusMetrics()
-	go serveMetrics(port)
-}
 
 func registerPrometheusMetrics() {
 	prometheus.MustRegister(requestsApproved)
@@ -75,18 +67,10 @@ func registerPrometheusMetrics() {
 	prometheus.MustRegister(requestsError)
 }
 
-func serveMetrics(port int) {
-	http.HandleFunc("/healthz", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("OK"))
-	})
-
-	http.Handle("/metrics", promhttp.Handler())
-
-	log.Fatal(http.ListenAndServe(":"+strconv.Itoa(port), nil))
-}
-
 func HandleRequests(filters inspectors.Inspectors, deniers inspectors.Inspectors, warners inspectors.Inspectors, deleteAfter time.Duration, client *kubernetes.Clientset) {
+	//Register prometheus metrics
+	registerPrometheusMetrics()
+
 	// Create a watcher and an informer for CertificateSigningRequests.
 	// The Add function
 	watchList := cache.NewListWatchFromClient(
